+++ guppy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.14.1] - 2022-03-18
+
+### Added
+
+- `Workspace::target_directory` returns the target directory provided in the Cargo metadata.
+- `Workspace::metadata_table` returns the freeform `workspace.metadata` table.
+
 ## [0.14.0] - 2022-03-14
 
 ### Added
@@ -493,6 +500,7 @@ lazy_static = "0.2"
 ### Added
 - Initial release.
 
+[0.14.1]: https://github.com/facebookincubator/cargo-guppy/releases/tag/guppy-0.14.1
 [0.14.0]: https://github.com/facebookincubator/cargo-guppy/releases/tag/guppy-0.14.0
 [0.13.0]: https://github.com/facebookincubator/cargo-guppy/releases/tag/guppy-0.13.0
 [0.12.6]: https://github.com/facebookincubator/cargo-guppy/releases/tag/guppy-0.12.6